requirements.txt
streamlit
gspread
authlib
google-auth
pandas
matplotlib
plotly
